IG-2557 wordTemplateBuilder created and functions moved out of emailGenerator
